[PM-1880] Implemented UpdateOrganizationUserGroupsCommand
