+++ src/Core/OrganizationFeatures/OrganizationUsers/Interfaces/IUpdateOrganizationUserGroupsCommand.cs
@@ -1,6 +1,8 @@
-﻿namespace Bit.Core.OrganizationFeatures.OrganizationUsers.Interfaces;
+﻿using Bit.Core.Entities;
 
-public class IUpdateOrganizationUserGroupsCommand
+namespace Bit.Core.OrganizationFeatures.OrganizationUsers.Interfaces;
+
+public interface IUpdateOrganizationUserGroupsCommand
 {
-
+    Task UpdateUserGroupsAsync(OrganizationUser organizationUser, IEnumerable<Guid> groupIds, Guid? loggedInUserId);
 }
